Update from /home/twfy-live/parl(data|parse) on mastodon
The knowledgeforge Subversion repository disappeared recently (the
Amazon instance it ran on had been accidentally shut down) and when it
was brought up again only had data up to 2010.  The previous commit in
my occasional mirror was from March 2014, and there had been quite a few
updates to parlparse and parldata since then.  This commit represents
taking the state of parlparse and parldata from mastodon as of 6th May
2014 and committing it.