Translate people XML data to JSON.
people.json now contains what was contained in the various XML files,
*members*.xml, peers-ucl.xml, royals.xml, people.xml, and *aliases*.xml.
personsets.py is no longer used.

The ministers-2010.json output is now sorted by person ID.

The new wrans parser outputs person IDs too, missed in 0124723c.

As aliases are now all in one file, the Ian Paisley alias is restricted
to be House of Commons, as they don't overlap in the Northern Ireland
Assembly and it would otherwise not know who to pick.