split Thomas Lewis' memberships over multiple terms
Thomas Lewis was listed as having a single membership from an 1828
general election to an 1834-12-29 general election. However, although
he _was_ actually an MP between those dates, it was not from a general
election, but a by-election (he resigned his existing Ennis seat to
contest Radnorshire upon the death of Walter Wilkins), and the
membership spans four parliamentary terms, which I’ve split out in
keeping with the approach used throughout of a single Membership for
each.

http://www.historyofparliamentonline.org/volume/1820-1832/member/lewis-t
homas-1780-1855