SP: update resolve names for new data
The fullnames data doesn't have the start_reason which we use as a belt
and braces check to make sure we have the presiding officer so fetch the
full member data and check that.

Use the new ministerial Popolo data for matching ministers. This data
uses the same end and start dates when chnages happen so in the case of
a match on a changeover day we pick the new minister.