SP: Update fetch script
If you request an official report ID that does not exist the site will
send you a 302 redirect to a search page. In that case we don't want to
store the page so we check the fetched URL against the requested on and
if they differ we don't store the page

Update URLs and ID matching.

Add option to fetch docs between two IDs. This is helpful for fetching a
set of old reports that have been missed for some reason. Also remove
the --historic option because this does the same.