SP: update parser for changes in HTML
There have been a few small changes to the HTML so update the parser to
cope with them. These include:

The HTML now wraps each speech in <p><span> which upsets the parser. It's
easier to just remove this wrapping than update the parser to deal with
it so do that.

In some question sessions the questions are numbered and the number
appears in front of the speaker name so we remove that to make sure the
matching works and add the number to the start of the next statement.

Handle a few extra official report titles

Some empty a tags appear with names which we can just skip.

We skip processing on empty speeches.

Some committee reports include lists which we just include verbatim.

If a page with an invalid ID has been requested then we get a search
page so check if we have on and bail out if we do. This should never
happen as the fetching code skips them but check to be sure.