Move and use ordayhd refs to spot major headings.
Allow through all <a name> from source HTML, not just those with no class or
class only of "anchor". This brings back many more semantic refs, such as
"subhd_0" or "ordayhd_1".

The issue with ordayhd is they appear after the heading, but we use the most
recent non-numeric reference before the heading as a guide. So move those refs
to before the heading first off, which lets us detect those headings as major
headings, improving the heading hierarchy.