Parse new hansard XML into publicwhip style XML
This can either be used as a module or as a command line script but
either way it gets instantiated with a file and a content type. It will
then take the contents of the XML and use the parser type specified
(debates, lords, westminhall or standing) to parse the Hansard XML and
output XML in the TWFY format.

It does this by fetching the root element for the debate type and then
interating over all the immediate children of the root and passing those
to parsing functions according to the tag name. It is mostly dumb about
context apart from a few places where it will skip a following tag in
order to join two headings together.

If it comes across a tag it does not recognize it will throw an
exception and processing will stop.

The first time it parses content for a date it will produce an 'a' file.
If it is subsequently handed content for that date it will generate a
'b' file and then compare the contents of the two. If they match then it
will delete the 'b' file, otherwise it will rewrite the 'a' file to set
up redirects from the 'a' content to the 'b' content.

The script itself does not track if it has seen an input file before so
if you hand it the same input file multiple times it will dumbly produce
and delete the 'b' file every time.

The output files are initially generated as temporary files and only
renamed at the end of processing. If the script dies half way through it
may leave the temporary file.

It's quite slow to start up because the member matching code loads and
processes the members.json file.

At the moment it does not work out the Hansard URLs for speeches etc.