improve capturing of speech text
use // as base for xpath expression so we get text inside child tags.

This means we have to remove any member tag from inside the speech as
otherwise we get the member's name at the start of the speech. Becuase
of the way lxml works the text after an element is stored as the tail of
that element so we need to assign that to the parent otherwise it's
removed with the element.

This does mean we're throwing away formatting at the moment but putting
that back in sensibly is a later task