Look at directories on Hansard XML files and run them through parser
This will look, by default, at today's set of XML fetched from Hansard
and process all the XML files in it that are for the Commons, Lords and
Public Bill Committees. It does this by matching the date in the
directory and then matching on the file name:

CHAN\d+.xml -> Commons
LHAN\d+.xml -> Lords
PBC.*xml -> Public Bill Committee

Commons files are passed to the parser twice, once for Commons debates
and once for Westminster Hall debates.

When a file has been succesfully processed its path and the content type
it was parsed for are written to a file. This list is loaded when the
script starts and if a file/content type is in there already it will not
be parsed again.

The script also accepts --from and --to options which take YYYY-MM-DD
dates as arguments to parse files from dates other than today.