special case Queens speech name handling
In order that TheyWorkForYou associates the Queen's speech with the
Queen we have to edit the XML by hand to add in a member for the speech.
As the Queen isn't a member of the Lords we need a special case to
recognise anyone with ContinuationText set to 'The Queen' as The Queen.