[UK] Make pims ID lookup date-specific.
This should help fix a long-standing issue where this code assumed a pims ID
was unique per person, when it is unique per office (e.g. all Bishops of
Oxford, or all (hereditary) Lords Reay, share a pims ID). By checking against
the date, we should never return the wrong person, or will error if we don't
yet have the new person in the system.